Meetings API separated, requirements.txt and .gitignore cleanup
