+++ meetings/delete_outdated_meetings.py
@@ -1,0 +1,5 @@
+from app import db
+from scripts import delete_outdated_meetings
+
+if __name__ == '__main__':
+    delete_outdated_meetings(db.session)
